Add claim section to realm panel
